+++ source/em/detail/header/data/HeaderData.cpp
@@ -4,11 +4,11 @@
 
 namespace em::detail::header {
     std::unordered_map<em::detail::header::DataType, unsigned int> byte_sizes = {
-        {em::detail::header::DataType::int8,    static_cast<unsigned int>(sizeof(int8_t))},
-        {em::detail::header::DataType::int16,   static_cast<unsigned int>(sizeof(int16_t))},
-        {em::detail::header::DataType::uint8,   static_cast<unsigned int>(sizeof(uint8_t))},
-        {em::detail::header::DataType::uint16,  static_cast<unsigned int>(sizeof(uint16_t))},
+        {em::detail::header::DataType::int8,    sizeof(int8_t)},
+        {em::detail::header::DataType::int16,   sizeof(int16_t)},
+        {em::detail::header::DataType::uint8,   sizeof(uint8_t)},
+        {em::detail::header::DataType::uint16,  sizeof(uint16_t)},
         {em::detail::header::DataType::float16, 2u},
-        {em::detail::header::DataType::float32, static_cast<unsigned int>(sizeof(float))},
+        {em::detail::header::DataType::float32, sizeof(float)},
     };
 }
